chore: Merge VocalWithMockedUseVocal test into Vocal.test.jsx
The standalone file existed only because vi.mock('../../hooks/useVocal',
...) is hoisted and would have contaminated the rest of Vocal.test.jsx
if added at the top level. Moves the test into Vocal.test.jsx using
vi.doMock + vi.resetModules + dynamic import('../Vocal') — the same
isolation pattern used by useCommands.test.js for its fuse.js fallback
test — and deletes the standalone file.

Closes #127.
